Realized I was creating completely disconnected dependencyValidators when appending, which is why I needed n*2 passes to detect the circular dependency when injected into fields (using Providers). The first pass would inflate only 1 level of the Provider<> dependencies, then the 2nd pass would connect the disconnected graphs and fail correctly.
When injecting multiple real objects, there is no single top-most point to the dependency graph.

This diff refactors DependencyValidator so it contains a Set<> of parents instead of just one. When appending, we no longer make new copies of the target’s children. Instead, we add ourselves as a new parent to the target’s children and validate all our new children against ourselves (and up).

This should have the extra benefit of being more efficient, because (a) we’re creating fewer DependencyValidator objects and (b) we’re no longer re-validating a tree that’s already been validated.