Realized i had an issue i didn’t like where shared resources weren’t properly combined with the test class when using a Mockspresso.Rule. This was because we’d build a complete mockspresso instance when building the rule (which included the shared resources) and then build upon it in the rule’s statement. This prevented us from “combining” the shared resources with the test class itself in this scenario.
This diff reverts to the original way we built the @Rule, by pass an instance of the MockspressoBuilderImpl that was used to build it, except we’ve add a single replacable field for testClass, so we don’t have the issue of continually adding to our array of test resources.

The question is, does this invalidate the last-mile setup/teardown code I just added?