Replace our typeToken with guava's via shadow jar
Not fully implemented yet, but looks promising. However there's a bug in IJ causing the references to be invisible inside the project (this should not be true when used externally though). Once we get the prototype working, we can ship this lib seperately and just depend on it as an external lib.