[BUG] Allow mockspresso to handle generic TypeVariable params by upgrading TypeToken (#4)
* Add tests to show bug using generic constructors

add tests for field and method injection

static objs

fix injector usage in test so we don't have to include the dagger dep

add test that provides a specific mock for the generic constructor param

* Replace our typeToken with guava's via shadow jar (#5)

* Replace our typeToken with guava's via shadow jar

Not fully implemented yet, but looks promising. However there's a bug in IJ causing the references to be invisible inside the project (this should not be true when used externally though). Once we get the prototype working, we can ship this lib seperately and just depend on it as an external lib.

* apidep on guavapack

* resolve types from context where appropriate in realObjectMaker. now im running into an odd index out of bounds issue when looping through constructor params

* update gdmc deps, everything still seems fine

* DO NOT COMMIT - just wanna see which builds passing on CI

* Revert "DO NOT COMMIT - just wanna see which builds passing on CI"

This reverts commit 63b1bb2f892df0ce5494a495e2ab8df41a4738c6.

* fix problems in test compilations. Tests are now all passing, we just have a frustrating compilation hickup that's breaking CI

* minimize guava, cuts out a decent chunk

* now that we're using a normal library, tests should start passing

* dont need shadow

* drop .tools

* use release version of mockspresso-reflect-guava, kill snapshot repo

* update changelog

* typo