[BREAKING] Remove Builder.specialObjectMakers(List) method (#16)
Removes a method from our Mockspresso.Builder api: Mockspreso.Builder.specialObjectMakers(List<SpecialObjectMaker).

This "pluralized" method is the only one of its kind in the api and has no good reason to exist. It is technically a breaking change, but a very easy one to fix. If there is a need to "group" multiple special object makers together, its recommended to use a plugin instead.